Stop reparenting dmenu
When I was embedding dmenu in a window before, it was actually a child
of that window, but that made the opacity and dim setting stop working
(because (I think) child windows are only as opaque as their parents).
SO I changed the rules so that dmenu no longer embeds itself in the
parent window, per se, but uses the window's geometry as bounds to its
geometry instead. That's all it really ever needed anyway. The behavior
is the same: -m is the flag, and it's either 0x<windowid> or <screenid>.